Refactor TcpSocketConnection related classes
Corrected misspelled name of class TcpServerOriginatedSocketConnection.
Splitted classes TcpSocketConnection and TcpServerOriginatedSocketConnection in two separate files.

Related task : APPLINK-30284
