+++ src/components/transport_manager/include/transport_manager/tcp/tcp_server_originated_socket_connection.h
@@ -1,4 +1,7 @@
 /*
+ * \file tcp_server_originated_socket_connection.h
+ * \brief TcpServerOriginatedSocketConnection class header file.
+ *
  * Copyright (c) 2017, Ford Motor Company
  * All rights reserved.
  *
@@ -30,52 +33,52 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "transport_manager/transport_adapter/transport_adapter_controller.h"
-#include "transport_manager/tcp/tcp_connection_factory.h"
-#include "transport_manager/tcp/tcp_server_originated_socket_connection.h"
+#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_TCP_SERVER_ORIGINATED_SOCKET_CONNECTION_H_
+#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_TCP_SERVER_ORIGINATED_SOCKET_CONNECTION_H_
 
-#include "utils/logger.h"
-#include "utils/make_shared.h"
+#include "transport_manager/transport_adapter/threaded_socket_connection.h"
 
 namespace transport_manager {
 namespace transport_adapter {
 
-CREATE_LOGGERPTR_GLOBAL(logger_, "TransportManager")
+class TransportAdapterController;
 
-TcpConnectionFactory::TcpConnectionFactory(
-    TransportAdapterController* controller)
-    : controller_(controller) {}
+/**
+ * @brief Class responsible for communication over sockets that originated by
+ * server.
+ */
+class TcpServerOriginatedSocketConnection : public ThreadedSocketConnection {
+ public:
+  /**
+   * @brief Constructor.
+   *
+   * @param device_uid Device unique identifier.
+   * @param app_handle Handle of application.
+   * @param controller Pointer to the device adapter controller.
+   */
+  TcpServerOriginatedSocketConnection(const DeviceUID& device_uid,
+                                      const ApplicationHandle& app_handle,
+                                      TransportAdapterController* controller);
 
-TransportAdapter::Error TcpConnectionFactory::Init() {
-  return TransportAdapter::OK;
-}
+  /**
+   * @brief Destructor.
+   */
+  virtual ~TcpServerOriginatedSocketConnection();
 
-TransportAdapter::Error TcpConnectionFactory::CreateConnection(
-    const DeviceUID& device_uid, const ApplicationHandle& app_handle) {
-  LOG4CXX_AUTO_TRACE(logger_);
-  LOG4CXX_DEBUG(logger_,
-                "DeviceUID: " << &device_uid
-                              << ", ApplicationHandle: " << &app_handle);
-  utils::SharedPtr<TcpServerOriginatedSocketConnection> connection =
-      utils::MakeShared<TcpServerOriginatedSocketConnection>(
-          device_uid, app_handle, controller_);
-  controller_->ConnectionCreated(connection, device_uid, app_handle);
-  const TransportAdapter::Error error = connection->Start();
-  if (TransportAdapter::OK != error) {
-    LOG4CXX_ERROR(logger_,
-                  "TCP ServerOriginated connection::Start() failed with error: "
-                      << error);
-  }
-  return error;
-}
-
-void TcpConnectionFactory::Terminate() {}
-
-bool TcpConnectionFactory::IsInitialised() const {
-  return true;
-}
-
-TcpConnectionFactory::~TcpConnectionFactory() {}
+ protected:
+  /**
+   * @brief Check if we can start the connection attempt and establish
+   *connection status.
+   *
+   * @param error contains information of any error that occurred during
+   *connection attempt.
+   *
+   * @return result that states whether we successfully connected or not.
+   */
+  virtual bool Establish(ConnectError** error);
+};
 
 }  // namespace transport_adapter
 }  // namespace transport_manager
+
+#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_TCP_SERVER_ORIGINATED_SOCKET_CONNECTION_H_
